+++ mail-mock/src/main/java/org/prebird/mailmock/service/MailRateLimiter.java
@@ -1,0 +1,5 @@
+package org.prebird.mailmock.service;
+
+public interface MailRateLimiter {
+  boolean limit(String account);
+}
